+++ spec/lib/fediverse/node_info_spec.rb
@@ -1,0 +1,64 @@
+require 'rails_helper'
+
+require 'fediverse/node_info'
+
+RSpec.describe Fediverse::NodeInfo do
+  describe '.fetch' do
+    let(:domain) { 'mamot.fr' }
+    let(:wk_nodeinfo_url) { "http://#{domain}/.well-known/nodeinfo" }
+
+    context 'when node is a valid activitypub node' do
+      it 'returns a Hash' do
+        VCR.use_cassette 'fediverse/nodeinfo/get_200' do
+          expect(described_class.fetch(domain)).to be_a Hash
+        end
+      end
+    end
+
+    context 'when node does not support activitypub' do
+      let(:nodeinfo_url) { 'https://mamot.fr/nodeinfo/2.0' }
+      let(:fake_wk_response) do
+        { 'links' => [{ 'rel' => 'http://nodeinfo.diaspora.software/ns/schema/2.0', 'href' => nodeinfo_url }] }
+      end
+      let(:fake_response) do
+        {
+          'version'           => '2.0',
+          'software'          => {
+            'name'    => 'mastodon',
+            'version' => '4.3.4',
+          },
+          'protocols'         => ['not_activitypub'],
+          'services'          => {
+            'outbound' => [],
+            'inbound'  => [],
+          },
+          'usage'             => {},
+          'openRegistrations' => true,
+          'metadata'          => {
+            'nodeName'        => 'Mamot - Le Mastodon de La Quadrature du Net ',
+            'nodeDescription' => 'Mamot.fr est un serveur Mastodon francophone, géré par La Quadrature du Net.',
+          },
+        }
+      end
+
+      before do
+        allow(Federails::Utils::JsonRequest).to receive(:get_json).with(wk_nodeinfo_url, follow_redirects: true).and_return(fake_wk_response).once
+        allow(Federails::Utils::JsonRequest).to receive(:get_json).with(nodeinfo_url).and_return(fake_response).once
+      end
+
+      it 'raises an exception' do
+        expect { described_class.fetch(domain) }.to raise_error Fediverse::NodeInfo::NoActivityPubError
+      end
+    end
+
+    context 'when node has no nodeinfo' do
+      before do
+        allow(Federails::Utils::JsonRequest).to receive(:get_json).with(wk_nodeinfo_url, follow_redirects: true).and_raise(Federails::Utils::JsonRequest::UnhandledResponseStatus).once
+      end
+
+      it 'raises an exception' do
+        expect { described_class.fetch(domain) }.to raise_error Federails::Utils::JsonRequest::UnhandledResponseStatus
+      end
+    end
+  end
+end